feat: Add colors.scss file and update package.json
This commit adds a new colors.scss file to the styles directory, which defines the primary color schema. It also updates the package.json file to include the sass dependency. These changes improve the styling capabilities of the application and enable the use of the primary color throughout the codebase.