feat: Add revalidate API endpoint
This commit adds a new API endpoint called revalidate, which is responsible for revalidating product data. It checks the authorization token and the query parameter to determine the type of data to revalidate. If the query parameter is "product", it triggers a revalidation process for the "/product/static" route. This addition enhances the application's data revalidation capabilities and improves the overall data integrity.